refactor: Update notification utility to use new notification library
